Create JOIN_IN_SUBQUERY.sql
wanted to use a combination of JOINs and WHERE to retrieve a selection of films that are in english and french languages while in horror and documentary genres. Ended up not working out and had to use AND condition in JOIN clause because WHERE requires the variable to be in both connecting tables (among other constraints such as not able to use for aggregate functions). Use cases of WHERE and HAVING would be researched further; rule of thumb is to use HAVING after GROUP BY and that GROUP BY is one of the last functions that specifies how temporary table should be retrieved before HAVING clause sorts the data in the table under some (aggregate) functions.

failed attempt is the bottom query while the successful attempt is the top query. Thanks to google and stackoverflow, I was reminded of what conditions I can use for JOINs while I'm still familiarizing myself with the documentations.